Add docs and further optimization to setup.py

+++ setup.py
@@ -1,13 +1,24 @@
 import subprocess
 import sys
-from sys import platform
+import os.path
 from shutil import copyfile
 
 def install(packages):
+    """ Invokes pip to install list of packages.
+
+        Parameters:
+            packages(list<str>): Packages to install, in requirements.txt format
+    """
     for package in packages:
         subprocess.call([sys.executable, "-m", "pip", "install", package])
 
 def getListFromFile(file, ignoreStart="#"):
+    """ Return list of packages to install given a requirements.txt file.
+
+        Parameters:
+            file(str): The requirements.txt file to parse.
+            ignoreStart(str): Ignore all lines that start with this string (optional).
+    """
     list = []
     with open(file) as handle:
         for line in handle.readlines():
@@ -16,21 +27,22 @@ def getListFromFile(file, ignoreStart="#"):
     return list
 
 
-# Get dependencies
+# Get dependencies for each OS
 requirements = getListFromFile("setup/requirements.txt")
 windows = getListFromFile("setup/windows.txt", "-r")
-linux = []
+linux = getListFromFile("setup/linux.txt", "-r")
 mac = getListFromFile("setup/mac.txt", "-r")
 
 # Install dependencies
 install(requirements)
-if platform == 'win32':
+if sys.platform == 'win32':
     install(windows)
-if platform.startswith('linux'):
+if sys.platform.startswith('linux'):
     install(linux)
-if platform == 'darwin':    # MacOS
+if sys.platform == 'darwin':    # MacOS
     install(mac)
 
-# Copy config template
-copyfile("./setup/config_template.ini", "config.ini")
-print("Config file created")
+# Create config file
+if not os.path.isfile("config.ini"):
+    copyfile("./setup/config_template.ini", "config.ini")
+    print("Config file created")
